requirements.txt
pygame
pyserial
pybluez2
ultralytics